+++ config/routes.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 RecordingStudioSupport::Engine.routes.draw do
+  resource :messages, only: :show, controller: "staff_messages"
+
   resources :sections, only: %i[index show new create edit update] do
     member do
       get :instant_search, to: "instant_searches#show"