+++ test/dummy/config/initializers/recording_studio_messages.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+RecordingStudioMessages.configure do |config|
+end