+++ test/dummy/config/initializers/recording_studio_notifications_email.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RecordingStudioNotificationsEmail.configure do |config|
+  config.from = "notifications@example.test"
+end